add page CabinetArea.js + Css file, rename BuyDomain page to  Register, create  Home.css and import to Home.js
