--- src/pages/BuyDomainForm.css
@@ -1,2 +1,0 @@
-.form {
-}
