+++ src/pages/Register.css
@@ -1,0 +1,47 @@
+div p {    
+    margin-left: 15px;
+    margin-right: 15px;
+}
+
+.row {
+    margin-left: 15px;
+    margin-right: 15px;
+}
+
+.form label {
+    display: inline-block;
+    min-width:150px;
+}
+
+.form {
+    background:#2396d5;    
+    display: flex;
+    position: relative;
+    width: 100%;
+    max-width: 550px;
+    min-width: 320px;
+    margin: auto;
+    color: white;    
+    text-transform: uppercase;
+}
+
+.form fieldset {
+    outline: none;
+    border: none;
+}
+
+.active {
+    /*display: block;
+    margin: auto;
+    padding-bottom: 100px;
+    padding-top: 50px;*/
+}
+
+.pagination {
+    display: block;
+    position: relative;   
+}
+
+.none {
+    display: none;
+}
